Added NRFX_TWI_FLAG_TX_NO_RESTART flag in NRF I2C driver
Added NRFX_TWI_FLAG_TX_NO_RESTART flag to the NRF I2C driver so it does not generate a restart conditon in case of burst write. Connected this flag to I2C_MSG_RESTART of zephyr

if I2C_MSG_RESTART is set then Restart will be generated else it will not be generated.